// dashboard/src/utils/hasura-api/generated/schemas/nativeQueryItem.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { NativeQueryItemArguments } from './nativeQueryItemArguments';
import type { NativeQueryItemType } from './nativeQueryItemType';
import type { NativeQueryRelationship } from './nativeQueryRelationship';

export interface NativeQueryItem {
  root_field_name: string;
  type?: NativeQueryItemType;
  arguments?: NativeQueryItemArguments;
  code: string;
  returns: string;
  comment?: string;
  object_relationships?: NativeQueryRelationship[];
  array_relationships?: NativeQueryRelationship[];
}